// src/types/enums/error-messages.ts
export enum ErrorMessages {
  requireName = 'Name is required. ',
  capitalizeError = 'Name should be capitalized',
  numbersInName = 'Name should not contain numbers',
  invalidFormat = 'Invalid format',
  requireAge = 'Age is required',
  positiveError = 'Age must be a positive number',
  requireEmail = 'Email is required',
  requireGender = 'Gender is required',
  requireAccept = 'Accept is required',
}